Fixes for OS X

+++ zsh_athame_setup.sh
@@ -99,11 +99,11 @@ if [ $redownload = 1 ]; then
 fi
 if [ ! -f zsh-5.1.1.tar.gz ]; then
   curl -O http://www.zsh.org/pub/old/zsh-5.1.1.tar.gz
-  if [ "$(md5sum zsh-5.1.1.tar.gz)" != "8ba28a9ef82e40c3a271602f18343b2f  zsh-5.1.1.tar.gz" ]; then
-    rm zsh-5.1.1.tar.gz
-    echo "FAILED: Incorrect md5 hash" >&2
-    exit 1
-  fi
+fi
+if [ "$(md5sum zsh-5.1.1.tar.gz 2>/dev/null)" != "8ba28a9ef82e40c3a271602f18343b2f  zsh-5.1.1.tar.gz" ] && [ "$(md5 zsh-5.1.1.tar.gz 2>/dev/null)" != "MD5 (zsh-5.1.1.tar.gz) = 8ba28a9ef82e40c3a271602f18343b2f" ]; then
+  rm zsh-5.1.1.tar.gz
+  echo "FAILED: Incorrect md5 hash" >&2
+  exit 1
 fi
 
 if [ ! -d zsh-5.1.1_tmp ]; then
